requirements.txt
# Core Dependencies
python-dateutil==2.8.2
pytz==2023.3

# Web Framework
streamlit==1.15.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Geospatial Libraries
gdal==3.4.3
rasterio==1.2.10
geopandas==0.10.2
shapely==2.0.2
fiona==1.9.5
pyproj==3.6.1
rtree==1.1.0

# Satellite Data API
sentinelhub==3.9.1
requests==2.31.0
oauthlib==3.2.2
requests-oauthlib==1.3.1

# Weather Data
pyowm==3.3.0

# Numerical Computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# Machine Learning / Deep Learning
tensorflow==2.8.4
keras==2.8.0
scikit-learn==1.0.2
joblib==1.3.2

# Image Processing
pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0

# Visualization
matplotlib==3.7.3
plotly==5.11.0
folium==0.12.1
branca==0.6.0
seaborn==0.13.0

# Database
sqlalchemy==2.0.23
geoalchemy2==0.14.2
sqlite3-python==1.0.0  # Built-in but listed for clarity

# Caching and Task Queue
redis==6.2.0
celery==5.2.7
kombu==5.3.4

# PDF Report Generation
reportlab==4.0.7
pypdf2==3.0.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
click==8.1.7
colorama==0.4.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# JSON/Data Handling
orjson==3.9.10
jsonschema==4.20.0

# Geospatial Web Services
owslib==0.29.3

# NetCDF for MOSDAC Data
netcdf4==1.6.5
h5py==3.10.0
xarray==2023.12.0

# Time Series Analysis
statsmodels==0.14.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code Quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Jupyter (for notebooks)
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1
ipywidgets==8.1.1

# Monitoring and Logging
loguru==0.7.2
sentry-sdk==1.39.1

# Authentication (for future API security)
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.1

# Email/SMS Notifications (optional)
# Uncomment if you plan to implement these features
# sendgrid==6.11.0
# twilio==8.11.0

# Cloud Storage (optional, for deployment)
# boto3==1.34.14  # AWS S3
# google-cloud-storage==2.14.0  # Google Cloud Storage
# azure-storage-blob==12.19.0  # Azure Blob Storage
